% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{Gompertz}
\alias{Gompertz}
\title{Gompertz}
\usage{
Gompertz(x, a = 1, b = 1, c = 1)
}
\arguments{
\item{x}{input vector}

\item{a}{see details}

\item{b}{see details}

\item{c}{see details}
}
\description{
maps numeric vector using Gompertz function
}

