% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{inverse_Gompertz}
\alias{inverse_Gompertz}
\title{Inverse Gompertz}
\usage{
inverse_Gompertz(x)
}
\arguments{
\item{x}{input vector Gompertz values}
}
\description{
maps numeric vector using Gompertz function
}
