% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{leakly_relu}
\alias{leakly_relu}
\alias{relu_leaky}
\title{Leaky Rectified Linear Unit}
\usage{
relu_leaky(x)
}
\arguments{
\item{x}{input vector}
}
\description{
maps numeric vector using leaky ReLU function
}
