% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{logistic}
\alias{logistic}
\title{Standard Logistic}
\usage{
logistic(x, k = 1, x0 = 0)
}
\arguments{
\item{x}{input vector}

\item{k}{see details}

\item{x0}{see details}
}
\description{
maps numeric vector using logistic function
}

