% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{relu_output_to_derivative}
\alias{relu_output_to_derivative}
\title{ReLU Derivative}
\usage{
relu_output_to_derivative(x)
}
\arguments{
\item{x}{vector or ReLU values}
}
\description{
Converts output of ReLU function to its derivative.
}
