% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{softplus_output_to_derivative}
\alias{softplus_output_to_derivative}
\title{SoftPlus Derivative}
\usage{
softplus_output_to_derivative(x)
}
\arguments{
\item{x}{vector of SoftPlus values}
}
\description{
Convert output of SoftPlus function to its derivative.
}
