% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{tanh_output_to_derivative}
\alias{tanh_output_to_derivative}
\title{Tanh Derivative}
\usage{
tanh_output_to_derivative(x)
}
\arguments{
\item{x}{vector of tanh values}
}
\description{
Convert output of tanh function to its derivative.
}

