% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid}
\usage{
sigmoid(x, method = c("logistic", "Gompertz"), inverse = FALSE,
  SoftMax = FALSE, k = 1, x0 = 0, a = 1, b = 1, c = 1)
}
\arguments{
\item{x}{numeric vector}
}
\description{
computes sigmoid nonlinearity
}

